Abstraction of COmmand
