+++ userConfig.txt
@@ -1,0 +1,1 @@
+Harris ,Male,18,1.7,71.0
